add empty progress bar & fix test
