+++ test/unit/optimizer/test_gp_operators.py
@@ -539,7 +539,8 @@ def test_no_opt_or_graph_nodes_after_mutation():
                                                          max_arity=3, max_depth=2, pop_size=5, num_of_generations=4,
                                                          crossover_prob=.8, mutation_prob=1)
     graph_params = get_pipeline_generation_params(composer_requirements, DEFAULT_DAG_RULES, task)
-    mutation = Mutation(mutation_types, graph_params, composer_requirements)
+    mutation = Mutation(mutation_types=mutation_types, graph_generation_params=graph_params,
+                        requirements=composer_requirements)
     new_graph, _ = mutation._adapt_and_apply_mutations(new_graph=graph, num_mut=1)
     new_pipeline = adapter.restore(new_graph)
